ai-service/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.4
python-multipart>=0.0.9
pydantic>=2.6.1
joblib>=1.3.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
google-generativeai>=0.4.0
